backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
requests==2.32.3

# Database & ORM
# SQLAlchemy dependencies removed - using MongoDB Atlas instead
# sqlalchemy==2.0.25
# alembic==1.13.1
# psycopg2-binary==2.9.9
# asyncpg==0.29.0
# aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# AI/ML Libraries (Optional - install separately if needed)
# torch>=2.6.0
# transformers>=4.36.0
# scikit-learn>=1.3.0
# numpy>=1.24.4
# pandas>=2.1.4

# Code Analysis
radon==6.0.1
pylint==3.0.3
psutil==5.9.6
memory-profiler==0.61.0
ast-tools==0.1.0

# Carbon Tracking
codecarbon==2.2.4
electricitymaps-api==0.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.39.2

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Rate Limiting
slowapi==0.1.9

# Email
fastapi-mail==1.4.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Report Generation
reportlab==4.0.7
openpyxl==3.1.2

# MongoDB Driver
motor==3.6.0
pymongo==4.10.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

